ladashy_unified/requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Existing dependencies
requests==2.31.0
paramiko==3.4.0
jinja2==3.1.2
markdown==3.5.1
pyyaml==6.0.1

# For desktop UI (optional)
customtkinter==5.2.1
pillow==10.2.0
